app/requirements.txt
streamlit==1.9.0
geopy==2.1.0
matplotlib==3.3.3
numpy==1.19.4
python-dotenv==0.18.0
google-generativeai==0.4.1
plotly==5.3.1
firebase-admin==4.5.0
